requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
ipykernel 
streamlit

fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
